The user control module was created
The user control module was created, adding the CRUD functions that compose this part of the system.
Additionally, it establishes the relationship of account creation associated with the login to manage user access.
Finally, it works simultaneously with the other modules present in the system, without errors in their interactions.