The cash control module was created
The cash control module was created, adding a punctual list that will allow to evaluate the statistics concerning the total income, total expenses and the total in general that the establishment generates.
Additionally, it works in conjunction with the other modules present in the system, without presenting errors between their interactions.