The purchase module was created
The purchase module was created, taking into account the main list that registers the movements, the section to select the necessary elements to order and finally the section that allows to specify the quantity of elements to buy.
On the other hand, the option to exit the administrator menu to return to the initial login has been adapted.
Finally, the built-in module works in conjunction with the other modules of the system.