The Vendor module was created
The salesman module was created, having its access from the main login. Having as main parts the inventory visualization and the sales management.
Additionally the option to return to the login from the salesman module was implemented.
Finally, this module works in conjunction with the other modules incorporated in the system, thus ensuring a complete structure.